Separate component scanner into two objects. Fix test
